Add files via upload
I am happy🌟to announce that I've completed the first task which is "Number Guessing Game" in c++ as a intern at Codsoft. 
This opportunity marks an exciting step in my career , where I'II be honing my skills in c++ programming, problem-solving and software developed. I'm eager to learn, grow and contribute to meaningful projects during this internship. Grateful to Codsoft for this opportunity, and looking forward to making the most of this experience.